1.4.3 is Images and Arrays
Images use 2 and 3 dimensional arrays, so images and arrays go hand in hand. There are 1, 2, and 3 dimensional arrays (3-dimensional arrays are arrays for color pictures). Pixels use an RGB form. You can call the first element for each list inside an array to derive channels. Finally, we review the class and packages.